Move the sources under the /src folder
Resolves: #1316
Change-type: patch
