+++ tests/util.spec.ts
@@ -2,17 +2,17 @@
 import * as _ from 'lodash';
 import { expect } from 'chai';
 import * as bSemver from 'balena-semver';
-import { Application } from '../lib';
+import { Application } from '../src';
 
 // HACK: Avoid typescript trying to resolve built es2015 files
 const nodeRequire = require;
 
 const { mergePineOptions } = nodeRequire(
 	'../es2015/util',
-) as typeof import('../lib/util');
+) as typeof import('../src/util');
 const { getDeviceOsSemverWithVariant } = nodeRequire(
 	'../es2015/util/device-os-version',
-) as typeof import('../lib/util/device-os-version');
+) as typeof import('../src/util/device-os-version');
 
 describe('Pine option merging', function () {
 	it('uses the defaults only, if no extra options are provided', function () {
